Use stack allocation instead of smart pointers.
The idea of using smart pointers in the list of actors only makes sense
if we use virtual inheritance, which seems unlikely at the moment.

Smart pointers wrapping around TCOD classes made sense with the idea
that the window size might change, but it won't, and grid.width and
.height expose how large they should be from the beginning.

In both cases, smart pointers just overcomplicate things which can be
done quite simply with things as they are.

Signed-off-by: Scott Prager <splinterofchaos@gmail.com>