Weapon slot A.
Instead of using an inventory index (II), just use an imaginary slot,
'A' (not 'a'). The II would become invalidated whenever an item was
consumed or dropped. With the weapon in its own slot, this never
happens.

This removes the need for Actor::weapon().
Change Actor::wpn to Actor::weapon.

Signed-off-by: Scott Prager <splinterofchaos@gmail.com>